refactor: Refactor code and edit interface
