+++ src/app/Helpers/RatingHelper.php
@@ -1,8 +1,13 @@
 <?php
+    use App\Models\Evaluates;
     // Calculator Rating of Products.
     if (!function_exists('getRatingProduct')) {
         function getRatingProduct($product_id)
         {
-            return round(\App\Models\Evaluates::where('product_id', $product_id)->avg('rating'), 1);
+            $stars = round(Evaluates::byProductId($product_id)->avg('rating'), 1);
+            if ($stars != 0) {
+              return $stars . "<i class='fa fa-star click-active' aria-hidden='true'></i>";
+            }
+            return __('custom.message_rating_no_data');
         }
     }
