+++ src/resources/views/home.blade.php
@@ -21,7 +21,7 @@
                     <div class="product-info text-center">
                       <img src="/storage/products/{{ $row->images->first()->image }}" alt="" />
                       <h2>{{ formatPrice($row->price) }}</h2>
-                      <p>{{ getRatingProduct($row->id) }} <i class="fa fa-star click-active" aria-hidden="true"></i></p>
+                      <p>{!! getRatingProduct($row->id)  !!}</p>
                       <p>{{ $row->name }}</p>
                       <a href="javascript:void(0)" class="btn btn-default add-to-cart" onclick="addToCart({!!$row->id!!},1,'add')">
                         <i class="fa fa-shopping-cart"></i>{{ __('custom.add_to_cart') }}
@@ -30,7 +30,7 @@
                     <div class="product-overlay">
                       <div class="overlay-content">
                         <h2>{{ formatPrice($row->price) }}</h2>
-                        <p>{{ getRatingProduct($row->id) }} <i class="fa fa-star click-active" aria-hidden="true"></i></p>
+                        <p>{!! getRatingProduct($row->id)  !!}</p>
                         <p><a href="{{ route('product_detail', ['slug' => $row->slug]) }}">{{ $row->name }}</a></p>
                         <a href="javascript:void(0)" class="btn btn-default add-to-cart" onclick="addToCart({!!$row->id!!},1,'add')">
                           <i class="fa fa-shopping-cart"></i>{{ __('custom.add_to_cart') }}
