+++ src/resources/views/web/carts/order-products.blade.php
@@ -59,13 +59,15 @@
           <li>{{__('custom.cart_total')}} <span id="cart-subtotal">{{ formatPrice(Cart::subtotal()) }}</span></li>
           <li>{{__('custom.eco_tax')}} <span>0đ</span></li>
           <li>{{__('custom.shipping_cost')}} <span>{{__('custom.free')}}</span></li>
-          <li>{{__('custom.total_price')}} <span id="cart-subtotal">{{ formatPrice(Cart::subtotal()) }}</span></li>
-          <li id="total_area_discount">
-            <input type="text" class="form-control" id="" value="" placeholder="{{__('custom.discount_code')}}">
-          </li>
+          <li>{{__('custom.total_price')}} <span id="cart-subtotal-total">{{ formatPrice(Cart::subtotal()) }}</span></li>
+          @if (Auth::check())
+            <li id="total_area_discount">
+              <input type="text" class="form-control" id="" value="" placeholder="{{__('custom.discount_code')}}">
+            </li>
+          @endif
         </ul>
         @if (Auth::check())
-          <a class="btn btn-danger check_out" id="cart_btn-order" href="javascript:void(0);">{{__('custom.order_product')}}</a>
+          <a class="btn btn-danger check_out" id="cart_btn-order" href="javascript:void(0)">{{__('custom.order_product')}}</a>
         @endif
       </div>
     </div>
